resolve error reponse bug in botcamps

+++ controllers/courses.js
@@ -10,13 +10,13 @@ const ErrorResponse = require('../utils/errorResponse')
 exports.getCourses = asyncHandler(async (req, res, next) => {
 
     if (req.params.bootcampId) {
-       const courses = await Course.find({ bootcamp: req.params.bootcampId })
+        const courses = await Course.find({ bootcamp: req.params.bootcampId })
 
-       return res.status(200).json({
-        success:true,
-        count: courses.length,
-        data: courses
-       })
+        return res.status(200).json({
+            success: true,
+            count: courses.length,
+            data: courses
+        })
     } else {
         res.status(200).json(res.advancedResults)
     }
@@ -32,33 +32,45 @@ exports.getCourses = asyncHandler(async (req, res, next) => {
 
 //get single course 
 //GET /api/v1/course/:id
-exports.getCourse = asyncHandler(async (req,res,next)=>{
-   const course = await Course.findById(req.params.id).populate({
-    path: 'bootcamp',
-    select: 'name description'
-   })
+exports.getCourse = asyncHandler(async (req, res, next) => {
+    const course = await Course.findById(req.params.id).populate({
+        path: 'bootcamp',
+        select: 'name description'
+    })
 
-   if(!course){
-    return next(new errorResponse(`No course with the id of ${req.params.id}`),404)
+    if (!course) {
+        return next(new errorResponse(`No course with the id of ${req.params.id}`), 404)
 
-   }
+    }
 
-   res.status(200).json({
-    success: true,
-    data: course
-   })
+    res.status(200).json({
+        success: true,
+        data: course
+    })
 })
 
 //add a course
 //POST /api/v1/bootcamps/:bootcampId/courses
 //private
-exports.addCourse = asyncHandler( async (req, res,next)=>{
+exports.addCourse = asyncHandler(async (req, res, next) => {
     req.body.bootcamp = req.params.bootcampId
+
+    req.body.user = req.user.id
 
     const bootcamp = await Bootcamp.findById(req.params.bootcampId)
 
-    if(!bootcamp){
-        return next(new ErrorResponse(`No bootcamp with the id of ${req.params.bootcampId}`),404)
+    if (!bootcamp) {
+        return next(new ErrorResponse(`No bootcamp with the id of ${req.params.bootcampId}`), 404)
+    }
+
+    // Make sure user is bootcamp owner
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(
+                `User ${req.user.id} is not authorized to add a course to bootcamp ${bootcamp.id}`,
+                401
+            )
+        );
     }
 
     const course = await Course.create(req.body)
@@ -72,19 +84,31 @@ exports.addCourse = asyncHandler( async (req, res,next)=>{
 //update courses
 // put /api/v1/courses/:id
 //private
-exports.updateCorse = asyncHandler( async (req,res,next)=>{
+exports.updateCorse = asyncHandler(async (req, res, next) => {
     let course = await Course.findById(req.params.id)
 
-    if(!course){
+    req.body.user = req.user.id
+
+    if (!course) {
         return next(
             new errorResponse(`No course with the id of ${re.params.id}`)
             , 404
         )
     }
 
-    course = await Course.findByIdAndUpdate(req.params.id, req.body,{
+    // Make sure user is course owner
+    if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(
+                `User ${req.user.id} is not authorized to update course ${course.id}`,
+                401
+            )
+        );
+    }
+    course = await Course.findByIdAndUpdate(req.params.id, req.body, {
         new: true, runValidators: true
-    })  
+    })
+
 
     res.status(200).json({
         success: true,
@@ -93,13 +117,26 @@ exports.updateCorse = asyncHandler( async (req,res,next)=>{
 })
 
 
-exports.deleteCourse = asyncHandler(async (req,res,next)=>{
-     const course = await Course.findById(req.params.id)
-    if(!course){
+exports.deleteCourse = asyncHandler(async (req, res, next) => {
+    const course = await Course.findById(req.params.id)
+
+    req.body.user = req.user.id
+
+
+    if (!course) {
         return next(
             new errorResponse(`No course with the id of ${re.params.id}`)
             , 404
         )
+    }
+    // Make sure user is course owner
+    if (course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(
+                `User ${req.user.id} is not authorized to delete course ${course.id}`,
+                401
+            )
+        );
     }
     await course.deleteOne()
     res.status(200).json({
